requirements.txt
apimatic-core==0.2.4
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.4
APScheduler==3.10.4
asgiref==3.6.0
black==23.7.0
Brotli==1.0.9
CacheControl==0.12.14
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
currencyapicom==0.1.1
deprecation==2.1.0
dj-database-url==1.2.0
Django==4.2.1
django-apscheduler==0.6.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.5
enum34==1.1.10
everapi==0.1.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gspread==5.10.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpickle==3.0.1
jsonpointer==2.4
MarkupSafe==2.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
paypalrestsdk==1.13.1
Pillow==10.0.0
platformdirs==3.9.1
protobuf==4.24.1
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.1.1
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
squareup==29.0.0.20230720
stripe==5.4.0
tomli==2.0.1
typing_extensions==4.6.1
tzlocal==5.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.2.0

